**Changelog:**
- Implement more options for the Inventory HUD (closes #173)
- Add Skin Manager (#176)
- Add profile importing & exporting
- Update to 1.21.9/10
- Fix some more bugs
- Preserve the current screen on config style change (#179)

+++ common/src/main/java/io/github/axolotlclient/modules/screenshotUtils/ScreenshotCopying.java
@@ -29,41 +29,57 @@ import java.awt.datatransfer.Transferable;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
+import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 
 import io.github.axolotlclient.AxolotlClientCommon;
-import lombok.AllArgsConstructor;
 import lombok.experimental.UtilityClass;
 import org.jetbrains.annotations.NotNull;
 import org.lwjgl.glfw.GLFW;
+import org.lwjgl.sdl.SDLClipboard;
+import org.lwjgl.sdl.SDL_ClipboardCleanupCallback;
+import org.lwjgl.sdl.SDL_ClipboardDataCallback;
+import org.lwjgl.system.MemoryStack;
+import org.lwjgl.system.MemoryUtil;
 
 @UtilityClass
 public class ScreenshotCopying {
 
 	static {
-		boolean wayland = false;
+		boolean sdl = false;
 		try {
-			wayland = GLFW.glfwGetPlatform() == GLFW.GLFW_PLATFORM_WAYLAND;
-		} catch (Throwable ignored) {
+			Class.forName("org.lwjgl.sdl.SDL");
+			sdl = true;
+		} catch (Throwable ignored) {}
+		SDL_AVAILABLE = sdl;
+		boolean wayland = false;
+		if (!sdl) {
+			try {
+				wayland = GLFW.glfwGetPlatform() == GLFW.GLFW_PLATFORM_WAYLAND;
+			} catch (Throwable ignored) {
+			}
 		}
 		IS_WAYLAND = wayland;
 	}
 
+	private static final boolean SDL_AVAILABLE;
 	private static final boolean IS_WAYLAND;
 
 	public void copy(Path file) {
-		if (IS_WAYLAND) {
-			copyWayland(file);
-		} else {
-			Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new FileTransferable(file.toFile()), null);
+		if (!SDL_AVAILABLE || !copySdl(file)) {
+			if (IS_WAYLAND) {
+				copyWayland(file);
+			} else {
+				Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new FileTransferable(file.toFile()), null);
+			}
 		}
 	}
 
 	private void copyWayland(Path f) {
 		try {
-			ProcessBuilder builder = new ProcessBuilder("bash", "-c", "wl-copy -t image/png < '" + f.toAbsolutePath()+"'");
+			ProcessBuilder builder = new ProcessBuilder("bash", "-c", "wl-copy -t image/png < '" + f.toAbsolutePath() + "'");
 			Process p = builder.start();
 			p.waitFor();
 		} catch (IOException | InterruptedException ignored) {
@@ -72,24 +88,50 @@ public class ScreenshotCopying {
 	}
 
 	public void copy(byte[] image) {
-		if (IS_WAYLAND) {
-			try {
-				Path i = Files.createTempFile("axolotlclient_screenshot", ".png");
-				Files.write(i, image);
-				copyWayland(i);
-				Files.delete(i);
-			} catch (IOException e) {
-				AxolotlClientCommon.getInstance().getLogger().error("Failed to copy image using temporary file!");
+		if (!SDL_AVAILABLE || !SDLFence.copySdl(image)) {
+			if (IS_WAYLAND) {
+				try {
+					Path i = Files.createTempFile("axolotlclient_screenshot", ".png");
+					Files.write(i, image);
+					copyWayland(i);
+					Files.delete(i);
+				} catch (IOException e) {
+					AxolotlClientCommon.getInstance().getLogger().error("Failed to copy image using temporary file!");
+				}
+			} else {
+				Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new ImageTransferable(image), null);
 			}
-		} else {
-			Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new ImageTransferable(image), null);
 		}
 	}
 
-	@AllArgsConstructor
-	protected static class ImageTransferable implements Transferable {
-		private final byte[] image;
+	private boolean copySdl(Path p) {
+		try {
+			return SDLFence.copySdl(Files.readAllBytes(p));
+		} catch (IOException e) {
+			AxolotlClientCommon.getInstance().getLogger().error("Failed to read screenshot!", e);
+		}
+		return false;
+	}
 
+	// Classloading fence...
+	static class SDLFence {
+		private static boolean copySdl(byte[] data) {
+			try (MemoryStack stack = MemoryStack.stackPush()) {
+				var mimetypesBuf = stack.UTF8("image/png");
+				var mimetypes = stack.pointers(mimetypesBuf);
+				var dataBuf = ByteBuffer.allocateDirect(data.length).put(data).flip();
+				var pointer = MemoryUtil.memAddress(dataBuf);
+				return SDLClipboard.SDL_SetClipboardData(SDL_ClipboardDataCallback.create((userdata, mime_type, size) -> {
+						var sizeBuf = MemoryUtil.memLongBuffer(size, 8);
+						sizeBuf.put(0, data.length);
+						return userdata;
+					}),
+					SDL_ClipboardCleanupCallback.create(userdata -> MemoryUtil.memFree(dataBuf)), pointer, mimetypes);
+			}
+		}
+	}
+
+	protected record ImageTransferable(byte[] image) implements Transferable {
 		@Override
 		public DataFlavor[] getTransferDataFlavors() {
 			return new DataFlavor[]{DataFlavor.imageFlavor};
@@ -107,10 +149,7 @@ public class ScreenshotCopying {
 		}
 	}
 
-	@AllArgsConstructor
-	protected static class FileTransferable implements Transferable {
-		private final File file;
-
+	protected record FileTransferable(File file) implements Transferable {
 		@Override
 		public DataFlavor[] getTransferDataFlavors() {
 			return new DataFlavor[]{DataFlavor.javaFileListFlavor};
@@ -122,7 +161,7 @@ public class ScreenshotCopying {
 		}
 
 		@Override
-		public Object getTransferData(DataFlavor flavor) {
+		public @NotNull Object getTransferData(DataFlavor flavor) {
 			final ArrayList<File> files = new ArrayList<>();
 			files.add(file);
 			return files;
